Log auction info contruction real time

+++ src/lib/GVD/src/Gvd.h
@@ -46,7 +46,7 @@ struct Gvd {
   GridGvd gridGvd;
   GvdGraph graphGvd;
   DistMap* distMap = NULL;
-  double updateTime = 0;
+  float updateTime = 0;
 
   // Info
   MapType& map;
